LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigate-event/intercept-handler-throws-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: a message

Harness Error (FAIL), message = Unhandled rejection: a message

PASS event.intercept() should abort if the handler throws
